+++ src/content/page.ts
@@ -50,7 +50,7 @@ const selectActiveElement = (port: Port, data: { cursor: number; cursorEnd: numb
 };
 
 // Given a full diff, apply it to INPUT and TEXTAREA elements directly.
-const applyDiff = (port: Port, data: any) => {
+const applyDiff = (port: Port, data: { cursor: number; cursorEnd: number; source: string }) => {
   if (document.activeElement) {
     const element = document.activeElement as HTMLElement;
     const cursorEnd = data.cursorEnd < data.cursor ? data.cursor : data.cursorEnd;
@@ -91,6 +91,55 @@ const copy = (port: Port, data: { text: string }) => {
   });
 };
 
+const findMatchAndScroll = (port: Port, data: { path: string }) => {
+  const path = data.path;
+  // Matches based on content ("path")
+  const snapshot = document.evaluate(
+    `.//*[not(self::script)][not(self::noscript)][not(self::title)][not(self::meta)][not(self::svg)][not(self::img)][not(self::style)][text()[contains(translate(., 'ABCDEFGHIJKLMNOPQRSTUVWXYZ', 'abcdefghijklmnopqrstuvwxyz'), '${path}')]]`,
+    document,
+    null,
+    XPathResult.ORDERED_NODE_SNAPSHOT_TYPE,
+    null
+  );
+  const matches = [];
+  for (let i = 0; i < snapshot.snapshotLength; i++) {
+    matches.push(snapshot.snapshotItem(i));
+  }
+  // A match that is below or to the right of the window
+  let target = matches.find((node) => {
+    const bounding = (node as Element).getBoundingClientRect();
+    return bounding.top >= window.innerHeight || bounding.left >= window.innerWidth;
+  });
+  // If no match, look for the first match that is above the window
+  if (target === undefined) {
+    target = matches.find((node) => {
+      const bounding = (node as Element).getBoundingClientRect();
+      return bounding.top < 0 || bounding.left < 0;
+    });
+  }
+  // If still no match (only match is in window), use the first one anyways
+  if (target === undefined && matches.length) {
+    target = matches[0];
+  }
+  // If still no match, stop
+  if (target === undefined) {
+    return;
+  }
+  const style = window.getComputedStyle(target as Element);
+  const backgroundColor = style.getPropertyValue("background-color");
+  (target as HTMLElement).style.backgroundColor = "yellow";
+  (target as HTMLElement).style.transition = "background-color 0.5s";
+  (target as Element).scrollIntoView({
+    block: "center",
+    inline: "center",
+    behavior: "smooth",
+  });
+  window.setTimeout(() => {
+    (target as HTMLElement).style.backgroundColor = backgroundColor;
+  }, 2000);
+  port.postMessage({ success: true });
+};
+
 chrome.runtime.onConnect.addListener((port) => {
   port.onMessage.addListener((msg) => {
     switch (msg.request) {
@@ -114,6 +163,9 @@ chrome.runtime.onConnect.addListener((port) => {
         break;
       case "setCursor":
         setCursor(port, msg.data);
+        break;
+      case "findMatchAndScroll":
+        findMatchAndScroll(port, msg.data);
         break;
       default:
         break;
